refactoring interceptor to internal package
