+++ sdk-workflows/src/main/java/io/dapr/workflows/client/DaprWorkflowClient.java
@@ -21,6 +21,7 @@ import io.dapr.client.Headers;
 import io.dapr.config.Properties;
 import io.dapr.utils.NetworkUtils;
 import io.dapr.workflows.Workflow;
+import io.dapr.workflows.internal.APITokenClientInterceptor;
 import io.grpc.CallOptions;
 import io.grpc.Channel;
 import io.grpc.ClientCall;
@@ -262,25 +263,6 @@ public class DaprWorkflowClient implements AutoCloseable {
     }
   }
 
-  public static ClientInterceptor WORKFLOW_INTERCEPTOR = new ClientInterceptor() {
-    @Override
-    public <ReqT, RespT> ClientCall<ReqT, RespT> interceptCall(
-        MethodDescriptor<ReqT, RespT> methodDescriptor,
-        CallOptions options,
-        Channel channel) {
-      // TBD: do we need timeout in workflow client?
-      ClientCall<ReqT, RespT> clientCall = channel.newCall(methodDescriptor, options);
-      return new ForwardingClientCall.SimpleForwardingClientCall<ReqT, RespT>(clientCall) {
-        @Override
-        public void start(final Listener<RespT> responseListener, final Metadata metadata) {
-          String daprApiToken = Properties.API_TOKEN.get();
-          if (daprApiToken != null) {
-            metadata.put(Metadata.Key.of(Headers.DAPR_API_TOKEN, Metadata.ASCII_STRING_MARSHALLER), daprApiToken);
-          }
-          super.start(responseListener, metadata);
-        }
-      };
-    }
-  };
+  private static ClientInterceptor WORKFLOW_INTERCEPTOR = new APITokenClientInterceptor();
 }
